+++ modal-dialog/src/main/java/raven/modal/option/LayoutOption.java
@@ -27,6 +27,10 @@ public class LayoutOption {
         return margin;
     }
 
+    public Insets getBackgroundPadding() {
+        return backgroundPadding;
+    }
+
     public DynamicSize getSize() {
         return size;
     }
@@ -51,9 +55,10 @@ public class LayoutOption {
         return onTop;
     }
 
-    private LayoutOption(DynamicSize location, Insets margin, DynamicSize size, DynamicSize animateDistance, RelativeToOwnerType relativeToOwnerType, boolean relativeToOwner, float animateScale, boolean onTop) {
+    private LayoutOption(DynamicSize location, Insets margin, Insets backgroundPadding, DynamicSize size, DynamicSize animateDistance, RelativeToOwnerType relativeToOwnerType, boolean relativeToOwner, float animateScale, boolean onTop) {
         this.location = location;
         this.margin = margin;
+        this.backgroundPadding = backgroundPadding;
         this.size = size;
         this.animateDistance = animateDistance;
         this.relativeToOwnerType = relativeToOwnerType;
@@ -69,6 +74,7 @@ public class LayoutOption {
     private Location horizontalLocation = Location.CENTER;
     private DynamicSize location = new DynamicSize(horizontalLocation.getValue(), Location.CENTER.getValue());
     private Insets margin = new Insets(7, 7, 7, 7);
+    private Insets backgroundPadding = new Insets(0, 0, 0, 0);
     private DynamicSize size = new DynamicSize(-1, -1);
     private DynamicSize animateDistance = new DynamicSize(0, 20);
     private RelativeToOwnerType relativeToOwnerType = RelativeToOwnerType.RELATIVE_CONTAINED;
@@ -101,6 +107,16 @@ public class LayoutOption {
 
     public LayoutOption setMargin(int margin) {
         this.margin = new Insets(margin, margin, margin, margin);
+        return this;
+    }
+
+    public LayoutOption setBackgroundPadding(int top, int left, int bottom, int right) {
+        this.backgroundPadding = new Insets(top, left, bottom, right);
+        return this;
+    }
+
+    public LayoutOption setBackgroundPadding(int adding) {
+        this.backgroundPadding = new Insets(adding, adding, adding, adding);
         return this;
     }
 
@@ -147,6 +163,10 @@ public class LayoutOption {
     }
 
     public LayoutOption copy() {
-        return new LayoutOption(location, new Insets(margin.top, margin.left, margin.bottom, margin.right), new DynamicSize(size), new DynamicSize(animateDistance), relativeToOwnerType, relativeToOwner, animateScale, onTop);
+        return new LayoutOption(location, copyInsets(margin), copyInsets(backgroundPadding), new DynamicSize(size), new DynamicSize(animateDistance), relativeToOwnerType, relativeToOwner, animateScale, onTop);
+    }
+
+    private Insets copyInsets(Insets insets) {
+        return new Insets(insets.top, insets.left, insets.bottom, insets.right);
     }
 }